support for remove item

+++ _views/_scenes/SceneEditorView.js
@@ -244,11 +244,37 @@ define(['jquery', 'backbone', 'fabric', 'BlockScene', 'BlockRSS', 'ScenesToolbar
                 self.m_canvas.discardActiveObject();
                 delete self.m_blocks[blockID];
                 pepper.removeScenePlayer(self.m_selectedSceneID, blockID);
-                BB.comBroker.fire(BB.EVENTS['SCENE_BLOCK_CHANGE'], self, null, null);
-                block.deleteBlock();
-                self._blockCountChanged();
+                self._loadScene();
+                // BB.comBroker.fire(BB.EVENTS['SCENE_BLOCK_CHANGE'], self, null, null);
+                // block.deleteBlock();
+                // self._blockCountChanged();
             });
         },
+
+        /**
+         Listen to when a user selects to delete a block
+         @method _listenSceneBlockRemove
+
+        _listenSceneBlockRemove: function () {
+            var self = this;
+            BB.comBroker.listen(BB.EVENTS.SCENE_ITEM_REMOVE, function () {
+                if (_.isUndefined(self.m_selectedSceneID))
+                    return;
+                var block = self.m_canvas.getActiveObject();
+                if (block == null)
+                    return;
+                var blockID = block.getBlockData().blockID;
+                self.m_canvas.discardActiveObject();
+                self.m_canvas.remove(block);
+                delete self.m_blocks[blockID];
+                pepper.removeScenePlayer(self.m_selectedSceneID, blockID);
+                // self._loadScene();
+                BB.comBroker.fire(BB.EVENTS['SCENE_BLOCK_CHANGE'], self, null, null);
+                block.deleteBlock();
+                // self._blockCountChanged();
+            });
+        },
+         **/
 
         /**
          Listen to when a user selects to delete a scene
@@ -852,8 +878,11 @@ define(['jquery', 'backbone', 'fabric', 'BlockScene', 'BlockRSS', 'ScenesToolbar
          **/
         _zoomReset: function () {
             var self = this;
-            if (_.isUndefined(self.m_selectedSceneID) || _.isUndefined(self.m_canvas))
+            if (_.isUndefined(self.m_selectedSceneID) || _.isUndefined(self.m_canvas)){
+                self.m_canvasScale = 1;
                 return;
+            }
+
             self.m_canvas.discardActiveGroup();
             self.m_canvas.setHeight(self.m_canvas.getHeight() * (1 / self.m_canvasScale));
             self.m_canvas.setWidth(self.m_canvas.getWidth() * (1 / self.m_canvasScale));
